pyproject.toml
[project]
name = "aseprite-mcp"
version = "0.1.0"
description = "A Python MCP server for interacting with the Aseprite API"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Divyansh Singh" }, { name = "Manh Luong" }]
dependencies = ["mcp>=1.6.0", "python-dotenv>=1.0.0"]

[dependency-groups]
dev = ["ruff>=0.9.0", "ty>=0.0.1"]
